+++ src/game/Object/PlayerMail.cpp
@@ -1,0 +1,224 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file PlayerMail.cpp
+ * @brief Cohesion split of Player.cpp.
+ *        Re-applied onto MangosOne TBC 2.4.3; same class, pure code move,
+ *        no behaviour change. CMake file(GLOB) picks this TU up automatically.
+ */
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "SpellMgr.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "CinematicFlyover.h"
+#include "SkillDiscovery.h"
+#include "QuestDef.h"
+#include "GossipDef.h"
+#include "UpdateData.h"
+#include "Channel.h"
+#include "ChannelMgr.h"
+#include "MapManager.h"
+#include "MapPersistentStateMgr.h"
+#include "InstanceData.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "Formulas.h"
+#include "Group.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Transports.h"
+#include "Weather.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGroundAV.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "ArenaTeam.h"
+#include "Chat.h"
+#include "revision_data.h"
+#include "Database/DatabaseImpl.h"
+#include "Spell.h"
+#include "ScriptMgr.h"
+#include "SocialMgr.h"
+#include "Mail.h"
+#include "DBCStores.h"
+#include "SQLStorages.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+#ifdef ENABLE_PLAYERBOTS
+#include "playerbot.h"
+#endif
+
+#include <cmath>
+
+#define ZONE_UPDATE_INTERVAL (1*IN_MILLISECONDS)
+
+#define PLAYER_SKILL_INDEX(x)       (PLAYER_SKILL_INFO_1_1 + ((x)*3))
+#define PLAYER_SKILL_VALUE_INDEX(x) (PLAYER_SKILL_INDEX(x)+1)
+#define PLAYER_SKILL_BONUS_INDEX(x) (PLAYER_SKILL_INDEX(x)+2)
+
+#define SKILL_VALUE(x)         PAIR32_LOPART(x)
+#define SKILL_MAX(x)           PAIR32_HIPART(x)
+#define MAKE_SKILL_VALUE(v, m) MAKE_PAIR32(v,m)
+
+#define SKILL_TEMP_BONUS(x)    int16(PAIR32_LOPART(x))
+#define SKILL_PERM_BONUS(x)    int16(PAIR32_HIPART(x))
+#define MAKE_SKILL_BONUS(t, p) MAKE_PAIR32(t,p)
+
+/**
+ * @brief Removes a mail entry from the player's in-memory mailbox.
+ *
+ * @param id The message identifier to remove.
+ */
+void Player::RemoveMail(uint32 id)
+{
+    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
+    {
+        if ((*itr)->messageID == id)
+        {
+            // do not delete item, because Player::removeMail() is called when returning mail to sender.
+            m_mail.erase(itr);
+            return;
+        }
+    }
+}
+
+/**
+ * @brief Sends the result of a mail operation to the client.
+ *
+ * @param mailId The mail message identifier.
+ * @param mailAction The mail action that was processed.
+ * @param mailError The result code for the action.
+ * @param equipError The equipment error code used for equip-related failures.
+ * @param item_guid The related item GUID low part.
+ * @param item_count The related item count.
+ */
+void Player::SendMailResult(uint32 mailId, MailResponseType mailAction, MailResponseResult mailError, uint32 equipError, uint32 item_guid, uint32 item_count)
+{
+    WorldPacket data(SMSG_SEND_MAIL_RESULT, (4 + 4 + 4 + (mailError == MAIL_ERR_EQUIP_ERROR ? 4 : (mailAction == MAIL_ITEM_TAKEN ? 4 + 4 : 0))));
+    data << (uint32) mailId;
+    data << (uint32) mailAction;
+    data << (uint32) mailError;
+    if (mailError == MAIL_ERR_EQUIP_ERROR)
+    {
+        data << (uint32) equipError;
+    }
+    else if (mailAction == MAIL_ITEM_TAKEN)
+    {
+        data << (uint32) item_guid;                         // item guid low?
+        data << (uint32) item_count;                        // item count?
+    }
+    GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Notifies the client that new mail is available.
+ */
+void Player::SendNewMail()
+{
+    // deliver undelivered mail
+    WorldPacket data(SMSG_RECEIVED_MAIL, 4);
+    data << (uint32) 0;
+    GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Recalculates unread mail count and the next pending delivery time.
+ */
+void Player::UpdateNextMailTimeAndUnreads()
+{
+    // calculate next delivery time (min. from non-delivered mails
+    // and recalculate unReadMail
+    time_t cTime = time(NULL);
+    m_nextMailDelivereTime = 0;
+    unReadMails = 0;
+    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
+    {
+        if ((*itr)->deliver_time > cTime)
+        {
+            if (!m_nextMailDelivereTime || m_nextMailDelivereTime > (*itr)->deliver_time)
+            {
+                m_nextMailDelivereTime = (*itr)->deliver_time;
+            }
+        }
+        else if (((*itr)->checked & MAIL_CHECK_MASK_READ) == 0)
+        {
+            ++unReadMails;
+        }
+    }
+}
+
+/**
+ * @brief Tracks a newly scheduled mail delivery time.
+ *
+ * @param deliver_time The time when the mail should be delivered.
+ */
+void Player::AddNewMailDeliverTime(time_t deliver_time)
+{
+    if (deliver_time <= time(NULL))                         // ready now
+    {
+        ++unReadMails;
+        SendNewMail();
+    }
+    else                                                    // not ready and no have ready mails
+    {
+        if (!m_nextMailDelivereTime || m_nextMailDelivereTime > deliver_time)
+        {
+            m_nextMailDelivereTime =  deliver_time;
+        }
+    }
+}
+
+/**
+ * @brief Finds a mail entry by message identifier.
+ *
+ * @param id The message identifier to search for.
+ * @return The matching mail entry, or null if none exists.
+ */
+Mail* Player::GetMail(uint32 id)
+{
+    for (PlayerMails::iterator itr = m_mail.begin(); itr != m_mail.end(); ++itr)
+    {
+        if ((*itr)->messageID == id)
+        {
+            return (*itr);
+        }
+    }
+    return NULL;
+}
